+++ src/main/java/org/jolokia/docker/maven/access/DockerAccess.java
@@ -7,7 +7,6 @@ import org.jolokia.docker.maven.access.log.LogCallback;
 import org.jolokia.docker.maven.access.log.LogGetHandle;
 import org.jolokia.docker.maven.config.Arguments;
 import org.jolokia.docker.maven.log.LogOutputSpec;
-import org.jolokia.docker.maven.model.*;
 import org.jolokia.docker.maven.model.Container;
 import org.jolokia.docker.maven.util.PomLabel;
 
@@ -101,16 +100,6 @@ public interface DockerAccess {
      */
     void stopContainer(String containerId, int killWait) throws DockerAccessException;
 
-    /**
-     * Get all containers matching a certain label. This might not be a cheap operation especially if many containers
-     * are running. Use with care.
-     *
-     * @param label label which the container must match
-     * @return list of container names matching the label.
-     * @see PomLabel#matches(PomLabel)
-     */
-    List<String> getContainersWithLabel(PomLabel label) throws DockerAccessException;
-
     /** Copy an archive (must be a tar) into a running container
      * Get all containers matching a certain label. This might not be a cheap operation especially if many containers
      * are running. Use with care.
@@ -123,7 +112,6 @@ public interface DockerAccess {
     void copyArchive(String containerId, File archive, String targetPath)
             throws DockerAccessException;
 
-     /**
     /**
      * Get logs for a container up to now synchronously.
      *